Extract API helpers to their own file with custom overrides
