+++ controllers/funding/materials-helpers.js
@@ -3,6 +3,7 @@ const { reduce } = require('lodash');
 const { check } = require('express-validator/check');
 
 const { errorTranslator } = require('../../modules/validators');
+const contentApi = require('../../services/content-api');
 
 const translateError = errorTranslator('global.forms');
 const translationLabelBase = 'funding.guidance.order-free-materials.formFields.';
@@ -239,8 +240,31 @@ If you have feedback, please contact matt.andrews@biglotteryfund.org.uk.`;
     return text.trim();
 }
 
+async function injectMerchandise(req, res, next) {
+    try {
+        res.locals.availableItems = await contentApi.getMerchandise(req.i18n.getLocale());
+        next();
+    } catch (error) {
+        next(error);
+    }
+}
+
+function injectMerchandiseCustom({ locale = false, showAll = false }) {
+    return async (req, res, next) => {
+        try {
+            const localeToUse = locale ? locale : req.i18n.getLocale();
+            res.locals.availableItems = await contentApi.getMerchandise(localeToUse, showAll);
+            next();
+        } catch (error) {
+            next(error);
+        }
+    };
+}
+
 module.exports = {
     materialFields,
     makeOrderText,
-    postcodeArea
+    postcodeArea,
+    injectMerchandise,
+    injectMerchandiseCustom
 };
